feat(casework): host reviews on the per-case page; search only navigates
Follow-up to the review redesign, based on real usage:

- The per-case page (/admin/reviews/case/:slug) now HOSTS the reviews: a run
  switcher plus the selected run's full breakdown inline (radar, sources,
  rule cards). Extracted that breakdown into a presentational
  ReviewResultView; the standalone per-run page is gone.
- /admin/reviews/:id is now a thin redirect to the case page
  (?run=<id>), so old links/bookmarks still resolve.
- Search (and list row-clicks) NAVIGATE to the case page instead of
  auto-starting a review. A review is triggered only from the case page's
  "Run review" button. Removed the per-row Re-run from the list.
- Selecting a run deep-links via ?run=; a never-reviewed case (now reachable
  via search) shows an empty state + the Run-review trigger.

Tests: rewrote the headless driver for the consolidated flow (search
navigates without submitting, case page hosts the switcher + inline
breakdown, Run-review triggers, legacy id redirects) — 20/20 checks pass.
