+++ tests/e2e/README.md
@@ -9,7 +9,7 @@ Drives the real SPA in headless Chromium (Playwright) against a mock backend tha
 - `admin-dates-repro.mjs` — admin case editor: BS/AD date pairing, the BS-picker corruption/crash path, bigo comma formatting, timeline inline inserts, and the save wire format.
 - `site-and-materials.mjs` — field-based material edit form (PUT body shape), case-detail evidence-title links to `/material/*`, home-page case-card thumbnail fallback.
 - `case-detail.mjs` — redesigned case page: banner (breadcrumb, badges, court-case @id IRI links), section jump nav, material-based evidence cards, PDF preview dialog, `/courtcase` page, `/case/<court-ref>` slug redirect, guest-chat-removal 404.
-- `casework-reviews.mjs` — casework review flow: autocomplete case-search submit (no raw-IRI box), the slimmed one-row-per-case list (latest run only), the per-case review page (`/admin/reviews/case/:slug`) run history, and the run-detail breakdown. Uses the in-memory review store the mock seeds (one case, two runs) and appends on submit.
+- `casework-reviews.mjs` — casework review flow: the slimmed one-row-per-case list (latest run only, navigational — no submit), search/row-click that navigate to the per-case page without starting a review, the per-case page (`/admin/reviews/case/:slug`) that HOSTS the run switcher + selected run's full breakdown inline and is the only place a review is triggered, and the legacy `/admin/reviews/:id` → `case/:slug?run=:id` redirect. Uses the in-memory review store the mock seeds (one case, two runs) and appends on submit.
 - `smoke.mjs` — quick home-page load check.
 
 ## Running
